refactored (moving populate-methods to own c files)
