+++ src/ui/nano/flow_user_confirm_outputs.h
@@ -1,0 +1,6 @@
+#ifndef SRC_UI_NANO_FLOW_USER_CONFIRM_OUTPUTS_H_
+#define SRC_UI_NANO_FLOW_USER_CONFIRM_OUTPUTS_H_
+
+void populate_data_outputs();
+
+#endif
